Remove old ops
